progress-monday

+++ src/controllers/cart.controllers.js
@@ -1,19 +1,56 @@
 const catchError = require('../utils/catchError');
 const Cart = require('../models/Cart');
+const Product = require('../models/Product');
+const Category = require('../models/Category');
+const ProductImg = require('../models/ProductImg');
 
 const getAll = catchError(async (req, res) => {
   const userId = req.user.id
   const results = await Cart.findAll({
     where: { userId },
     include: [
+      // Product
       {
         model: Product,
-        attributes: { exclude: ["updateAt", "createAt"] },
-        include: {
-          model: Catego
-        }
+        attributes: { exclude: ["updatedAt", "createdAt"] },
+        include: [
+          {
+            model: Category,
+            attributes: ['name']
+          },
+          {
+            model: ProductImg
+          }
+        ]
       }
     ]
+
+  });
+  return res.json(results);
+});
+
+const getOne = catchError(async (req, res) => {
+  const { id } = req.params
+  const userId = req.user.id
+  const results = await Cart.findByPk(id, {
+    where: { userId },
+    include: [
+      // Product
+      {
+        model: Product,
+        attributes: { exclude: ["updatedAt", "createdAt"] },
+        include: [
+          {
+            model: Category,
+            attributes: ['name']
+          },
+          {
+            model: ProductImg
+          }
+        ],
+      }
+    ]
+
   });
   return res.json(results);
 });
@@ -28,15 +65,21 @@ const create = catchError(async (req, res) => {
 
 const remove = catchError(async (req, res) => {
   const { id } = req.params;
-  await Cart.destroy({ where: { id }});
+  const userId = req.user.id
+  const result = await Cart.destroy({ where: { id, userId } });
+  if (!result) return res.sendStatus(404)
   return res.sendStatus(204);
 });
 
 const update = catchError(async (req, res) => {
+  const userId = req.user.id
   const { id } = req.params;
+  const { quantity } = req.body
+
+
   const result = await Cart.update(
-    req.body,
-    { where: { id }, returning: true }
+    { quantity },
+    { where: { id, userId }, returning: true }
   );
   if (result[0] === 0) return res.sendStatus(404);
   return res.json(result[1][0]);
@@ -44,6 +87,7 @@ const update = catchError(async (req, res) => {
 
 module.exports = {
   getAll,
+  getOne,
   create,
   remove,
   update
